+++ src/meta/brstm.c
@@ -62,7 +62,7 @@ VGMSTREAM * init_vgmstream_brstm(const char * const filename) {
 
     /* fill in the vital statistics */
     vgmstream->num_samples = read_32bitBE(head_offset+0x2c,infile);
-    vgmstream->sample_rate = read_16bitBE(head_offset+0x24,infile);
+    vgmstream->sample_rate = (uint16_t)read_16bitBE(head_offset+0x24,infile);
     /* channels and loop flag are set by allocate_vgmstream */
     vgmstream->loop_start_sample = read_32bitBE(head_offset+0x28,infile);
     vgmstream->loop_end_sample = vgmstream->num_samples;
